+++ app/src/main/java/com/eflc/mintdrop/room/Converters.kt
@@ -1,0 +1,16 @@
+package com.eflc.mintdrop.room
+
+import androidx.room.TypeConverter
+import java.time.LocalDateTime
+
+class Converters {
+    @TypeConverter
+    fun fromTimestamp(value: String?): LocalDateTime? {
+        return value?.let { LocalDateTime.parse(it) }
+    }
+
+    @TypeConverter
+    fun dateToTimestamp(date: LocalDateTime?): String? {
+        return date?.toString()
+    }
+}
