+++ app/src/main/java/com/eflc/mintdrop/room/dao/CategoryDao.kt
@@ -1,9 +1,9 @@
-package com.eflc.mintdrop.repository.dao
+package com.eflc.mintdrop.room.dao
 
 import androidx.room.Dao
 import androidx.room.Query
 import androidx.room.Upsert
-import com.eflc.mintdrop.repository.dao.entity.Category
+import com.eflc.mintdrop.room.dao.entity.Category
 import kotlinx.coroutines.flow.Flow
 
 @Dao
