+++ app/src/main/java/com/eflc/mintdrop/room/dao/entity/Category.kt
@@ -1,9 +1,9 @@
-package com.eflc.mintdrop.repository.dao.entity
+package com.eflc.mintdrop.room.dao.entity
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import java.util.Date
+import java.time.LocalDateTime
 
 @Entity(tableName = "category")
 data class Category(
@@ -16,7 +16,7 @@ data class Category(
     @ColumnInfo(name = "icon_ref")
     val iconRef: String,
     @ColumnInfo(name = "created_on")
-    val createdOn: Date,
+    val createdOn: LocalDateTime = LocalDateTime.now(),
     @ColumnInfo(name = "last_modified")
-    val lastModified: Date?
+    val lastModified: LocalDateTime?
 )
