+++ app/src/main/java/com/eflc/mintdrop/room/dao/entity/EntryHistory.kt
@@ -1,10 +1,10 @@
-package com.eflc.mintdrop.repository.dao.entity
+package com.eflc.mintdrop.room.dao.entity
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.PrimaryKey
-import java.util.Date
+import java.time.LocalDateTime
 
 @Entity(tableName = "entry_history", foreignKeys = [
     ForeignKey(
@@ -25,9 +25,9 @@ data class EntryHistory(
     @ColumnInfo(name = "description")
     val description: String,
     @ColumnInfo(name = "date")
-    val date: Date,
+    val date: LocalDateTime = LocalDateTime.now(),
     @ColumnInfo(name = "created_on")
-    val createdOn: Date,
+    val createdOn: LocalDateTime = LocalDateTime.now(),
     @ColumnInfo(name = "last_modified")
-    val lastModified: Date?
+    val lastModified: LocalDateTime?
 )
