+++ app/src/main/java/com/eflc/mintdrop/room/dao/entity/Subcategory.kt
@@ -1,10 +1,10 @@
-package com.eflc.mintdrop.repository.dao.entity
+package com.eflc.mintdrop.room.dao.entity
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.PrimaryKey
-import java.util.Date
+import java.time.LocalDateTime
 
 @Entity(tableName = "subcategory", foreignKeys = [
     ForeignKey(
@@ -27,9 +27,9 @@ data class Subcategory(
     @ColumnInfo(name = "icon_ref")
     val iconRef: String,
     @ColumnInfo(name = "last_entry_on")
-    val lastEntryOn: Date?,
+    val lastEntryOn: LocalDateTime?,
     @ColumnInfo(name = "created_on")
-    val createdOn: Date,
+    val createdOn: LocalDateTime = LocalDateTime.now(),
     @ColumnInfo(name = "last_modified")
-    val lastModified: Date?
+    val lastModified: LocalDateTime?
 )
